Moved all the functions of arraytools.py into the example plotepiscan.py
