+++ src/biotite/sequence/graphics/alignment.py
@@ -1110,38 +1110,23 @@ def get_colorbar(axes, array1, array2, colormap, transform ='linear',  orient =N
         return r'${}\cdot10^{{{}}}$'.format(a, b)
     
     if method == 'linear':
-        vmiA = df1['combined_signal'].min()
-        vmiB = df2['combined_signal'].min()
-        vmxA = df1['combined_signal'].max()
-        vmxB = df2['combined_signal'].max()
+        vmiA = df1['comb_signal'].min()
+        vmiB = df2['comb_signal'].min()
+        vmxA = df1['comb_signal'].max()
+        vmxB = df2['comb_signal'].max()
         # Colormap normalization:
         norm = mpl.colors.PowerNorm(gamma = 1.0, 
                                     vmin = min(vmiA,vmiB), vmax = max(vmxA,vmxB))
 
-    elif method == 'sqrt':
-        vmiA = df1['combined_signal'].min()
-        vmiB = df2['combined_signal'].min()
-        vmxA = df1['combined_signal'].max()
-        vmxB = df2['combined_signal'].max()
-        # Colormap normalization:
-        norm = mpl.colors.PowerNorm(gamma = 0.5, 
-                                    vmin = min(vmiA,vmiB), vmax = max(vmxA,vmxB))
     elif method == 'cubic':
-        vmiA = df1['combined_signal'].min()
-        vmiB = df2['combined_signal'].min()
-        vmxA = df1['combined_signal'].max()
-        vmxB = df2['combined_signal'].max()
+        vmiA = df1['comb_signal'].min()
+        vmiB = df2['comb_signal'].min()
+        vmxA = df1['comb_signal'].max()
+        vmxB = df2['comb_signal'].max()
         # Colormap normalization:
         norm = mpl.colors.PowerNorm(gamma = 0.33, 
                                     vmin = min(vmiA,vmiB), vmax = max(vmxA,vmxB))
-    elif method == 'log':
-        vmiA = df1['combined_signal'].min()
-        vmiB = df2['combined_signal'].min()
-        vmxA = df1['combined_signal'].max()
-        vmxB = df2['combined_signal'].max()
-        # Colormap normalization:
-        norm = mpl.colors.LogNorm(vmin = 1 if vmiA==0 or vmiB==0 else min(vmiA,vmiB),
-                                  vmax = max(vmxA,vmxB))
+
         
     fig = mpl.pyplot.figure()        
     return fig.colorbar(mpl.cm.ScalarMappable(norm = norm, cmap = cmp),
